Không dùng được HOC nên dùng tạm cái này
