+++ frontend/TH-ebook/package.json
@@ -32,6 +32,7 @@
     "react-redux": "^9.1.2",
     "react-router-dom": "^6.27.0",
     "redux": "^5.0.1",
+    "redux-logger": "^3.0.6",
     "redux-saga": "^1.3.0",
     "swiper": "^11.1.14",
     "tailwindcss": "^3.4.14",
